LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.lineTo.nonfinite.worker-expected.txt


PASS lineTo() with Infinity/NaN is ignored 
